security authentication

+++ src/main/java/com/yasas/hotel/model/RoomModel.java
@@ -2,12 +2,14 @@ package com.yasas.hotel.model;
 
 import lombok.*;
 
+import java.util.UUID;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class RoomModel {
 
-    private  Integer roomId;
+    private Long roomId;
 
     private String type;
