+++ src/main/java/com/yasas/hotel/repository/RoomRepository.java
@@ -4,9 +4,11 @@ import com.yasas.hotel.entity.RoomEntity;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.UUID;
+
 
 @Repository
-public interface RoomRepository  extends CrudRepository<RoomEntity, Integer> {
+public interface RoomRepository  extends CrudRepository<RoomEntity, Long> {
 
 
     Iterable<RoomEntity> findAllByType(String type);
